Обработчик открытия фото вынесен в отдельную функцию, удален лишний обработчик зарытия по клику
